Add files via upload
Created a program that generates "n" number of processes and performs matrix creation and multiplication.

Implemented the process creation logic by generating a random number between 10 and 99 and multiplying it by the last digit of the roll number. Divided the roll number by the generated number and took the modulo with 25. Added 15 if the number was less than 10.
Created a (n x n) matrix and populated it randomly.
Saved the matrix in a text file.
Each process performed matrix multiplication for one row and stored the result in the respective text file along with its ID.
The main process read all the text files of the processes and displayed the resultant matrix.
Please note that you should replace "n" in the commit description with the actual value of "n" that was used in your program.